실드 개편

+++ Assets/Script/ScriptSO/Player/RangedSO.cs
@@ -8,7 +8,7 @@ public class RangedUnitSO : UnitSO
     [Header("RangedUnitSO")]
     public int launchVolume;        // 한번의 발사의 발사량
     public int bulletSpread;        // 탄퍼짐
-    public int bulletLifeTime;      // 총알 유지시간
+    public float bulletLifeTime;      // 총알 유지시간
     public Sprite BulletSprite;      // 총알 스프라이트
 }
